requirements.txt
pytest>=7.0
allure-pytest==2.14.2
httpx>=0.28.1
jsonschema>=4.23.0
pydantic>=2.11.4
pydantic-settings>=2.9.1
loguru>=0.7.3